+++ glide/estimators/stratified_ppi.py
@@ -3,7 +3,7 @@ from typing import List, Tuple
 import numpy as np
 from numpy.typing import NDArray
 
-from glide.core.clt_confidence_interval import CLTConfidenceInterval
+from glide.confidence_intervals import CLTConfidenceInterval
 from glide.core.mean_inference_result import SemiSupervisedMeanInferenceResult
 from glide.core.utils import compute_effective_sample_size
 
@@ -206,13 +206,13 @@ class StratifiedPPIMeanEstimator:
 
         std = np.sqrt(weighted_var)
         n_true = np.sum(~np.isnan(y_true))
-        effective_sample_size = compute_effective_sample_size(y_true, std)
 
         confidence_interval = CLTConfidenceInterval(
             mean=weighted_mean,
             std=std,
             confidence_level=confidence_level,
         )
+        effective_sample_size = compute_effective_sample_size(y_true, confidence_interval.var)
         result = SemiSupervisedMeanInferenceResult(
             confidence_interval=confidence_interval,
             metric_name=metric_name,